requirements.txt
asn1crypto==0.22.0
beautifulsoup4==4.6.0
boto3==1.5.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
configparser==3.5.0
contextlib2==0.5.5
coverage==4.4.1
cryptography==1.9
enum34==1.1.6
flake8==3.3.0
funcsigs==1.0.2
html5lib==0.999999999
idna==2.5
ipaddress==1.0.18
mccabe==0.6.1
mock==2.0.0
pbr==3.1.0
pluggy==0.4.0
py==1.4.34
pycodestyle==2.3.1
pycparser==2.17
pyfiglet==0.7.5
pyflakes==1.5.0
pyOpenSSL==17.0.0
pytest==3.1.2
pytest-cov==2.5.1
PyYAML==3.12
requests==2.18.1
six==1.10.0
tox==2.7.0
urllib3==1.21.1
vcrpy==1.11.1
virtualenv==15.1.0
webencodings==0.5.1
websocket-client==0.44.0
wrapt==1.10.10
protobuf==3.0.0
tensorflow==1.4.0
h5py==2.7.1
pillow==2.1.0
keras==2.0.9
urlextract==0.7
